test(rules): split the tiling-family parameter tests out of test_ruleaction
test_ruleaction.cpp had grown to 1371 lines, past CLAUDE.md's 1150 hard
ceiling. The autotile, gap, tab-indicator and scrolling parameter suites
share a shape the rest of the file does not: each is a value-keyed action
whose entire contract is a numeric range or a closed token vocabulary.
Moving those four to their own target leaves the original owning the JSON
contract and the action-domain pins, at 934 and 469 lines.

Also: the tab-indicator style setter now has emit-count coverage. Its
re-seed is hand-written, so the two-key write firing settingsChanged once
and a same-value write firing nothing were both unpinned, and the leg
that stood in for the latter asserted a width that could not move either
way. Two QML comments corrected where they overstated what the code does.
