+++ libs/phosphor-rules/tests/CMakeLists.txt
@@ -43,6 +43,7 @@ pwr_add_test(phosphorrules_test_matchtypes test_matchtypes.cpp)
 pwr_add_test(phosphorrules_test_windowquery test_windowquery.cpp)
 pwr_add_test(phosphorrules_test_matchexpression test_matchexpression.cpp)
 pwr_add_test(phosphorrules_test_ruleaction test_ruleaction.cpp)
+pwr_add_test(phosphorrules_test_ruleaction_tilingparams test_ruleaction_tilingparams.cpp)
 pwr_add_test(phosphorrules_test_actionregistry test_actionregistry.cpp)
 pwr_add_test(phosphorrules_test_rule test_rule.cpp)
 pwr_add_test(phosphorrules_test_ruleset test_ruleset.cpp)
